refactor(lib): use instance uniq id composable - export id value directly
